Fix form kit dependencies
Make sure that simple form will be available in the production mode
of the playbook web app. This is needed so we can show the example of
how to use playbook with simple form. The previous code only had a
development dependency defined in the gemspec, moving this to the
Gemfile will make it available in the production web app.