Implement the valid_contacts method
Currently, the kit wasn't displaying the wrong number caption, nor documenting the functionality at all in the kit. This documents the occurance of a wrong number in a contact in Playbook, detects valid contacts, displays them above a wrong number caption and displays the wrong numbers beneath with the correct icon. This in no way is a backward breaking change to the Nitro repo, as it isn't implemented there yet.