Add textarea kit to form kit
Tickets: [BOP-309]

Extend the form builder for the form kit to add in a text area.
Previously the form builder used the default text area method provided
by ActionView::Helpers::FormBuilder and SimpleForm::FormBuilder. This
new version uses the text area element built buy these implementations,
but wraps them using the pb_textarea kit.

Generalize the text input builder to allow a kit name to be passed in.
This allows us to reuse the builder for the text_area field as well as
the existing text type inputs fields.