Display wrong contacts correctly (#499)
* Create documentation for wrong numbers

* Update contact type to more accurately describe itself

This describes the contact type of a wrong phone, differentiating between a wrong email, if added in the future.

* Implement the valid_contacts method

Currently, the kit wasn't displaying the wrong number caption, nor documenting the functionality at all in the kit. This documents the occurance of a wrong number in a contact in Playbook, detects valid contacts, displays them above a wrong number caption and displays the wrong numbers beneath with the correct icon. This in no way is a backward breaking change to the Nitro repo, as it isn't implemented there yet.

* Update changelog

* Add test coverage for wrong phone icon

* Additional test coverage for contact type icon

The example will default to the phone icon if passed anything other than the list of defined icons

* Add padding to the top of wrong number caption

* Move changelog entry to correct position

* Actually move the changelog entry to the correct position

For real this time

* Clean up extra lines

* Be more descriptive in changelog

* Use SCSS file instead of in line

* Pass the classname as a prop instead of a div wrap

Change the conditional to be an if statement instead