Add pb typeahead kit
For some Rails kits, we need small sprinkles of JavaScript to push
enhance the elements we use to build the kit. PbEnhancedElement
provides a easy way to active JavaScript for elements. Mutation
observer monitors the DOM for matching elements, and then creates a
new instance of the PbEnhancedElement and connects it. When the
element is removed from the DOM, the mutation observer disconnects the
element.

The expectation is that a user would subclass PbEnhancedElement, and
provide to sprinkle in JavaScript enhancements to the kit. A subclass
must provide either a selector static property, or redefine the
matches function. Additionally, it is assumed a custom connect function
will be defined.

This typeahead is built this the idea of decoupling suggestions from
the finding and selecting options. This makes it the job of other code
to provide the kit with options that match the users query.

By default a custom event is emitted when the user types into the kits
search input. This event can be listed to, and used to provide the
user with results for the query.

A custom event is also emitted when a selection has been made. This
event can be listed to and used to know a user has made a selection,
and then take further action.

A simple cache is used to improve performance and make it easier to
deal with out of order responses for the typeahead searches.

Update app/pb_kits/playbook/pb_typeahead/_typeahead.scss

Co-Authored-By: Stephen Marshall <smarshall1980@gmail.com>

Update app/pb_kits/playbook/pb_typeahead/typeahead.rb

Co-Authored-By: Stephen Marshall <smarshall1980@gmail.com>