Add a context option for the typeahead
Tickets: [BOP-357]

Allow the typeahead to use a context to make searches distinct when
using the same typeahead for the to provide results with different
conditions.

The typeahead has a cache, but that might not be the best name for
it. This structure was primarily added to provide coordination for
asynchronous searches for results. We can get the benefits of a cache,
and always provide the user with fresh results. We change the search
to always show the cached values when results are present but also
always issuing a event to get new results.