Release 4.4.0 (#630)
* Create date range stacked kit (#559)

* Upgrade ci-kubed for bugfixes (#627)

See https://github.com/powerhome/ci-kubed/pull/29

This played out on https://github.com/powerhome/playbook/pull/621 where the merge commit f2b0c81 had its image build reported against the previous commit 092039f in the CI job at https://ci.powerhrg.com/blue/organizations/jenkins/powerhome%2Fplaybook/detail/PR-621/6/pipeline/15

* Add a context option for the typeahead (#621)

* Add a context option for the typeahead

Tickets: [BOP-357]

Allow the typeahead to use a context to make searches distinct when
using the same typeahead for the to provide results with different
conditions.

The typeahead has a cache, but that might not be the best name for
it. This structure was primarily added to provide coordination for
asynchronous searches for results. We can get the benefits of a cache,
and always provide the user with fresh results. We change the search
to always show the cached values when results are present but also
always issuing a event to get new results.

* Correct popover offset docs (#611)

* Use mutation observer to bind behavior of PbPopover (#626)

* Fixed stat change typo in yml file (#629)

* Fixed stat change typo in yml file

* Updated changelog with stat change yml typo fix

* Doc Examples with Conditional Rendering (#623)

* Prepping Release

* Update CHANGELOG.md

Co-authored-by: Anil Cherukuri <anil.cherukuri03@gmail.com>
Co-authored-by: Ben Langfeld <ben@langfeld.me>
Co-authored-by: Terry Finn <terryfinn@users.noreply.github.com>
Co-authored-by: Stephen Marshall <smarshall1980@gmail.com>
Co-authored-by: Aaron Kuehler <aaron.kuehler@gmail.com>
Co-authored-by: Christina Tai <42459486+christinaatai@users.noreply.github.com>
Co-authored-by: Jasper Furniss  <jasperfurniss@gmail.com>