Support disabled aria-disabled for buttons
HTML buttons have a disabled attribute, and this prevents the button
form being clicked and focused. Additionally, buttons can have a
aria-disabled attrubte to indicate to a user the button is disabled. The
presence of either of these should render the same, but button is not
intended to be clickable.

For backwards compatibility, we are allowing the explicit presence of
the disabled classes to do the same thing, and adding the disabled or
aria-disabled attribute, but it would be better to encourage the use of
the disabled attributes to do this.