Make button kit compatible with Turbo/Hotwire features in Rails 7
Having the target attribute present on a link makes using the
`turbo_method` and `turbo_confirmation` features in Rails 7
incompatible. This PR ensures that a target attribute is only added
when `new_window` is explicitly specified in the props. This
additionally brings better parity between the Rails and React version
of the kits, as the React version does not currently set a value for
the target attribute unless `newWindow` is `true`.