[PLAY-746] Fixing Phone Number Input kit width and method errors (#2462)
**What does this PR do?**
- Removing the max-width limitation on the input that causes layout and
wiggle (when the selected country is changed) issues.
- Removing Rails onChange and isValid props since they are not used on
the Rails side and they are causing errors since they are null and we
are trying to call these functions.

**How to test?** Steps to confirm the desired behavior:
1. Go to Phone Number Input Kit
2. Check the Rails version if we are not getting errors when changing
the country or typing a number


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.

---------

Co-authored-by: Nida Ghuman <nidaqg@gmail.com>