[PLAY-788] Fix React Global Props Table (#2513)
**What does this PR do?** A clear and concise description with your
runway ticket url.

This PR is a clone of https://github.com/powerhome/playbook/pull/2474.
It had to be reverted due to a large number of issues. This PR had a
tricky merge conflict to resolve so this PR is the entirety of the
original PR and the merge conflict resolution

**Screenshots:** Screenshots to visualize your addition/change

![screenshot-127 0 0 1_3000-2023 05
01-16_07_29](https://user-images.githubusercontent.com/38965626/235522150-a488537f-5eb7-48ce-8d1d-2b4463e58781.png)


**How to test?** Steps to confirm the desired behavior:
1. Go to a kit on `playbook-website`
2. Click on any kit (make sure react)
3. Scroll down to the bottom of the page 
4. See the props table 


#### Checklist:
- [ ] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [😈] **TESTS** I have added test coverage to my code.

---------

Co-authored-by: Nida Ghuman <nidaqg@gmail.com>