[PBNTR-12] Adding check to empty filters (#2573)
**What does this PR do?** 
We are replacing the `<h4>` in the PB Filter kit with a `<span>` tag and
we are checking for empty filters in the Rails version since the current
version is not checking for all the empty values the filters prop could
have.

[Runway
Ticket](https://nitro.powerhrg.com/runway/backlog_items/PBNTR-12)


**How to test?** Steps to confirm the desired behavior:
1. Go to the Filters kit page in PB.


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.