Fix `flexGrow={0}`, `flexShrink={0}`
Setting the global props `flexGrow` and `flexShrink` to `0` doesn't work
because they are emitted using a truthiness check, which `0` fails.

This results in `flexGrow={0}` returning an empty string and not adding
the `flex_grow_0` class and consequently not applying the `flex-grow:
0;' style to the component.

This commit fixes the bug by replacing the truthiness check with an
inclusion check using the `Binary` type of the `flexGrow` and
`flexShrink` props.

Signed-off-by: Jasper <jasperfurniss@gmail.com>