Fix `flexGrow={0}`, `flexShrink={0}` (#2634)
**What does this PR do?** A clear and concise description with your
runway ticket url.

Setting the global props `flexGrow` and `flexShrink` to `0` doesn't work
because they are emitted using a truthiness check, which `0` fails.

This results in `flexGrow={0}` returning an empty string and not adding
the `flex_grow_0` class and consequently not applying the `flex-grow:
0;'
style to the component.

This commit fixes the bug by replacing the truthiness check with an
inclusion
check using the `Binary` type of the `flexGrow` and `flexShrink`
props.

**Screenshots:** Screenshots to visualize your addition/change

N/A

**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [x] **TESTS** I have added test coverage to my code.

Signed-off-by: Jasper <jasperfurniss@gmail.com>
Co-authored-by: Jasper Furniss <jasper.furniss@powerhrg.com>