Requesting More CPU (#2758)
**What does this PR do?** 

Paired with Aaron K. today and this is the result. 
We investigated our CPU resources on grafana, and came to conclusion
that for now, we can temporarily request an increase for production env
only. Staging is currently sufficient, and we aren't hitting any
ceilings there.

We verified that the templates will render by testing locally! 🎉 


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.