No CPU Limits + Increase Memory in Production (#2787)
I made these changes during Create and it seemed to eliminate pod
restarts, which are the cause of the 500 errors.

I believe there is some request queueing going on with Puma, and the
constrained resources don't allow the queue to be worked fast enough. I
recommend implementing some puma metrics to analyze the queue. But these
changes should help a bunch.