[PLAY-1432] Playground Sanitization Security (#3624)
**What does this PR do?** A clear and concise description with your
runway ticket url.

Runway https://runway.powerhrg.com/backlog_items/PLAY-1432

Anyone was able to input whatever erb into the playground. This was bad
because people could do shell scrips and other stuff by submitting
something like this `<%= system(“pwd”) %>`

The plan is to evaluate the ERB code coming in and if it has a method
that isn’t on the white list then the request is not processed

These methods have to be on the white list because the Parser parses
ruby and theses methods are needed to “run” erb

```
<<
To_s
+@
Freeze
```


More details about my thoughts here
https://huddle.powerapp.cloud/v2/projects/647/artifacts/6881