[PLAY-1592] Show htmlOptions as a Global Prop  (#3755)
**What does this PR do?** A clear and concise description with your
runway ticket url.
[PLAY-1592](https://runway.powerhrg.com/backlog_items/PLAY-1592)

As a Playbook user, I want the htmlOptions prop to display in Playbook
as a global prop instead of a kit prop (which is false) so that there is
no confusion to what type of prop htmlOptions is.

**Screenshots:** Screenshots to visualize your addition/change


**How to test?** Steps to confirm the desired behavior:
1. Go to any kit page
2. Ensure you are in the react tab
3. Scroll down to the available props list
4. The "htmlOptions" prop should not appear under the kit props list,
but will show under the global props list.


#### Checklist:
- [X] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [X] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.

Co-authored-by: Jasper Furniss <jasper.furniss@powerhrg.com>