[PLAY-1603] Add color hover docs and fix background hover (#3874)
**What does this PR do?** A clear and concise description with your
runway ticket url.
runway https://runway.powerhrg.com/backlog_items/PLAY-1603

Add a "Text Color" hover prop option is the ask in this story 

because color is already a hover option i just used that. The Background
and color hover props weren't working because a lot of our kits are
listening for classnames that are "_error", "_success" or
"_#{any_color_token}"

so i changed the classname to not trigger the pre existing css by making
the syntax "hover_color-#{any_color_token}". Doing this to the
background kit fixes it too

Checkout the page here
https://pr3874.playbook.beta.gm.powerapp.cloud/visual_guidelines/hover


![screenshot-127_0_0_1_3000-2024_11_01-09_29_40](https://github.com/user-attachments/assets/c3346cb5-019a-401f-9bbb-b3e45901c08f)