[PLAY-1647] Remove react modal from drawer  (#3891)
**What does this PR do?** A clear and concise description with your
runway ticket url.
Runway https://runway.powerhrg.com/backlog_items/PLAY-1647

Background: The drawer kit is currently in draft mode. Its similar to
the dialog kit, and i used a lot of the same code to make the drawer
kit.

This story removes `react-modal` from the drawer kit. It didn't do as
much as i thought it did and we can get away with just using a `div`

**How to test?** Steps to confirm the desired behavior:
1. Go to
https://pr3891.playbook.beta.px.powerapp.cloud/kits/drawer/react
2. click all the examples