[PBNTR-830] Advanced Table kit: Sticky Header & Responsive Prop working together - Rails (#4293)
**What does this PR do?** A clear and concise description with your
runway ticket url.
[PBNTR-830](https://runway.powerhrg.com/backlog_items/PBNTR-830) permits
a rails Advanced Table to have both a sticky header + responsiveness.
This is a followup to
[PBNTR-834](https://runway.powerhrg.com/backlog_items/PBNTR-834) which
gives this ability to the React version of the table.
This PR mimics the rails side exactly: the global max_height prop is
used as a kit prop to create a container within which the header is
sticky, and this is demonstrated in the Table Props Sticky Header doc
example. Additionally, the header in large screen mode is sticky within
the container - not sticky to the page but if rows are expanded past the
given max height it will get an overflow-y scroll bar. Meanwhile, the
preexisting Table Props doc example has been updated to non-responsive
and shows a sticky header that is sticky to the page itself.


**Screenshots:** Screenshots to visualize your addition/change
<img width="757" alt="for PR table props doc ex"
src="https://github.com/user-attachments/assets/02181c01-336a-47ee-a9b3-a581d078e6fc"
/>
<img width="1307" alt="for PR sticky header large window"
src="https://github.com/user-attachments/assets/d78983a1-85c8-436a-b811-fb6fa906de32"
/>
<img width="753" alt="for PR sticky header responsive window"
src="https://github.com/user-attachments/assets/6d9e958b-da69-4b09-b277-6bbb3b798d0e"
/>


**How to test?** Steps to confirm the desired behavior:
1. Go to the [Table Props doc example](link to come) and scroll down the
page to see the sticky header in action for a nonresponsive table. Do
this on a large window and a smaller one.
2. Go to the [Table Props Sticky Header doc example](link to come).
Expand at least 3 rows to activate the overflow-y based on the max
height given to the table. On a large window, scroll up and down and see
the sticky header for the container in action. Make the window smaller
and see the sticky header and responsiveness first sticky column both
work within the container of the table.


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
~~- [ ] **TESTS** I have added test coverage to my code.~~