[PBNTR-883] Advanced Table: Multi Column Headers to work with Custom Cells (#4292)
[Support ticket](https://runway.powerhrg.com/backlog_items/PBNTR-883)

When you tried to use the customRenderer to do custom cells while also
using the multi-column functionality, the custom cells did not work as
expected for the first column in each column group. The issue was that
the first column in each column group was being treated like the forst
column in the table and getting those styles. This fixes for that.

![Screenshot 2025-02-19 at 11 12
53 AM](https://github.com/user-attachments/assets/6be8eafd-6956-4370-87f2-5b14595a3260)