[PLAY-1922] Tooltip Interaction using @floating-ui (#4322)
**What does this PR do?** A clear and concise description with your
runway ticket url.

This PR adds the floating-ui library and uses it to add interaction
functionality to our tooltip in rails.
This is the same library that we use for react.
PopperJs is still used for all other tooltips. Only when
interaction=true is passed will floating-ui work.
A user will be able to keep the tooltip open and copy content out of it.

**Screenshots:** Screenshots to visualize your addition/change
![Screenshot 2025-02-25 at 12 11
02 PM](https://github.com/user-attachments/assets/23ee22a7-1fef-4de9-8cfa-202251fb3e8c)


**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.

---------

Co-authored-by: markdoeswork <mark@does.work>