[PLAY-1957] Fix Inline Date Picker Icon Logic (#4416)
I'm now realizing I may have taken this too far based on the stories
expected behavior. I made it so the + shows back up after removing the
selected date instead of staying a chevron. If we want to just have it
be a + all the time, I can stash this and I think it's a quick fix for
that. Let me know.

I also noticed that the sass was targeting these by id, which was fine
when only one input was on the page, but with multiple inputs I was
running into issues, so I changed that up a bit.
 
[PLAY-1957](https://runway.powerhrg.com/backlog_items/PLAY-1957)

[ENV](https://pr4416.playbook.beta.px.powerapp.cloud/kits/date_picker/react)