[PLAY-1950] Playbook Website: Fix Incorrect Links on Getting Started Page (#4432)
This fixes the links on the Getting Started Page. There is other logic
on the page to account for missing token pages and missing global pages.
None of that is being triggered currently on the page but I left it as
is for now. Let me know if I should remove it.

[PLAY-1950](https://runway.powerhrg.com/backlog_items/PLAY-1950)
[Review ENV](https://pr4432.playbook.beta.px.powerapp.cloud/)