[PLAY-1964] Table Kit: Fix Text Transform in Table Headers (#4452)
I think the simplest way to make the table headers respect the detail
and body kit styles, while not breaking any of the current table styles,
is to add a <code>text-transform: none</code> to them, while keeping the
<code>text-transform</code> on the table headers.

If we don't want to edit the detail and body kits at all, I can figure
out a different way to do it. Let me know.

REVISED: Taking a different approach here and only applying the current
styles if there isn't a nested playbook kit in the th element.

[PLAY-1964](https://runway.powerhrg.com/backlog_items/PLAY-1964)
[CSB](https://codesandbox.io/p/devbox/naughty-paper-52h4t4)

---------

Co-authored-by: Nida Ghuman <nidaqg@gmail.com>