[PLAY-1499] - Background Kit: Add Overlay w/ Opacity to Background Image (#4478)
[PLAY-1499](https://runway.powerhrg.com/backlog_items/PLAY-1499)

Here's a
[CSB](https://codesandbox.io/p/sandbox/busy-leavitt-4zxgsq?file=%2Fsrc%2FApp.js%3A10%2C32)
as well as the [Review
ENV](https://pr4478.playbook.beta.px.powerapp.cloud/kits/background/react)
with the React example. I added the example to the top of the page for
ease of review of the POC, but I can move it further down the page.
opacity_1 through opacity_10 can be used and all the colors that are
already available. Let me know if anything else is needed or if I should
take this further.