[PLAY-1906] - Overlay Kit: Add Opacity & Gradient prop - React (#4501)
[PLAY-1906](https://runway.powerhrg.com/backlog_items/PLAY-1906)

[CSB](https://codesandbox.io/p/devbox/keen-murdock-dkpvkv?workspaceId=ws_Ni9WCwV5jk56bH746ZU8Ee)
[Review
ENV](https://pr4501.playbook.beta.gm.powerapp.cloud/kits/overlay/react)

**EDITS: I added new example docs and split up the copy in the existing
default example.**

This PR adds support for a `gradient={false}` prop to the Overlay kit,
as well as an `opacity` prop for rendering a solid color overlay.
Support for `black`, `white`, `success`, and `error` colors has also
been added.

The original request called for a limited set of opacity values, but it
was easier to allow the full range.

I added copy to the example docs—let me know if a new example section
would be better. I was on the fence about that.

Tests for the new props are also included.
![Screenshot 2025-04-14 at 3 10
01 PM](https://github.com/user-attachments/assets/18981282-1e24-4725-bf2a-4b098429906f)
![Screenshot 2025-04-14 at 3 10
07 PM](https://github.com/user-attachments/assets/e81b68ef-b0d0-4e5c-8f86-6d8172bbc139)