[PLAY-2040] Remove Unnecessary Sticky Left/Right Column Classes in React Tables (#4528)
**What does this PR do?**
- When using classnames, in React, check that the stickyLeftColumn and
stickyRightColumn arrays have a length whether the class is added
- The default value is an empty array, `stickyLeftColumn = [],`, but
this is truthy so it was being added to every table.

https://runway.powerhrg.com/backlog_items/PLAY-2040

**Screenshots:**
![Screenshot 2025-04-21 at 2 35
42 PM](https://github.com/user-attachments/assets/a2a56cf1-5480-4f84-b9a4-0b71f8f4b31d)

**How to test?** Steps to confirm the desired behavior:
1. Go to Table and Advanced Table React kit docs
2. Inspect a majority of the docs, there should be no
`sticky-left-column` or `sticky-right-column` classes for the `<table>`
element
3. On /kits/table/react#sticky-left-column , `sticky-left-column` should
be a class on the `<table>` element
4. On /kits/table/react#sticky-right-column , `sticky-right-column`
should be a class on the `<table>` element
5. On /kits/advanced_table/react#sticky-columns, there should be no
`sticky-left-column` or `sticky-right-column` classes for the `<table>`
element. But, there should be an `advanced-table-sticky-left-columns` on
the wrapping div

#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.