[PLAY-1965] Overlay Kit: Add a "Full Screen" prop - React (#4537)
**What does this PR do?** A clear and concise description with your
runway ticket url.
[Runway Story](https://runway.powerhrg.com/backlog_items/PLAY-1965)

As a Playbook user,
I want to to be able to customize the color and opacity of a non
gradient Overlay,
So that I can ensure clarity and legibility of text when over images.

**Screenshots:** Screenshots to visualize your addition/change
![Screenshot 2025-04-22 at 1 34
34 PM](https://github.com/user-attachments/assets/91154cb7-8269-4190-b2ce-80a284d2b24e)


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.