[PLAY-2042] Popover Kit: Add Turbo Support - Rails (#4551)
[PLAY-2042](https://runway.powerhrg.com/backlog_items/PLAY-2042)

**EDIT** Added a new example doc to the [Filter
Kit](https://pr4551.playbook.beta.gm.powerapp.cloud/kits/filter/rails)

This PR adds an `append_to` prop on the rails side that allows the user
to choose what element the popover is appended to. By default it appends
to the `body` tag, but the user can pass `"parent"` to append to the
parent, or pass the class or id of any element.

Test [here](pb_popover_tooltip) by inspecting the DOM and see where the
pb_popover_tooltip is appended.