[PLAY-2129] Advanced Table: ColumnVisibility Control (#4579)
**What does this PR do?** 
[PLAY-2129](https://runway.powerhrg.com/backlog_items/PLAY-2129)

**This PR covers the following:**
Adds `columnVisibilityControl` prop that allows you to toggle visibility
of different columns off and on.
Allows devs to manage state for this control if they want so they can
set an initial state, etc
allows devs to customize which columns to show as options within the
dropdown so they can have a smaller dropdown if they want
Allows for visibility control to work with multi-header columns
For this one, the dropdown shows hierarchy of columns
unchecking and checking cascades up and down similar to MLS

 **To be covered in a follow up:**
With the responsiveness set to true, the dropdown is clipped at the
bottom where the table ends on smaller screen sizes (only an issue if
dropdown exceeds table height)
This is due the way overflow is set up, and will need to be looked at
separately

**Screenshots:**

![Screenshot 2025-05-08 at 11 50
37 AM](https://github.com/user-attachments/assets/bf415c63-ec4e-43fc-8968-d87fcfe0826e)

![Screenshot 2025-05-08 at 11 50
27 AM](https://github.com/user-attachments/assets/43d2ef72-5c3a-4031-8640-8519b1262ec7)


#### Checklist:
- [ ] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [ ] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.