[PLAY-2235] Advanced Table: Pinned Rows to always pin Children (#4747)
**What does this PR do?** A clear and concise description with your
runway ticket url.

[Runway Story](https://runway.powerhrg.com/backlog_items/PLAY-2235)

Row Pinning functionality was shipped on
[PLAY-2178](https://runway.powerhrg.com/backlog_items/PLAY-2178), for
this story we want to add further enhancements. Currently you have to
manually give the prop the id for all child rows if a parent in pinned.
We want to make it so if a parent is pinned, all children rows will
automatically be pinned as well.

**Screenshots:** Screenshots to visualize your addition/change


https://github.com/user-attachments/assets/491d78a2-3737-41b2-8ccc-28c9d561fa94

**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change


#### Checklist:
- [ ] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [ ] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.
- [ ] **SEMVER** I have added a `minor`, `major`, or `patch` label for
release.
- [ ] **RC** I have added an `inactive RC` label if not an active RC.