[PLAY-2255] Testing load time with refactored enhanced element file (#4754)
**What does this PR do?**

[PLAY-2255](https://runway.powerhrg.com/backlog_items/PLAY-2255)

Refactored the JS for the advanced Table to reduce the number of
querySelectAll calls + reduced the number of event listeners per table +
reduced the code around the action bar

Load time is mostly on the first render now, due to the logic in the rb
file creating all the rows. We may need to look into a lazy loading
piece for the website itself.


**Screenshots:** 


**How to test?** Steps to confirm the desired behavior:
1. Go to '...'
2. Click on '....'
3. Scroll down to '....'
4. See addition/change


#### Checklist:
- [x] **LABELS** Add a label: `enhancement`, `bug`, `improvement`, `new
kit`, `deprecated`, or `breaking`. See [Changelog &
Labels](https://github.com/powerhome/playbook/wiki/Changelog-&-Labels)
for details.
- [x] **DEPLOY** I have added the `milano` label to show I'm ready for a
review.
- [ ] **TESTS** I have added test coverage to my code.
- [x] **SEMVER** I have added a `minor`, `major`, or `patch` label for
release.
- [ ] **RC** I have added an `inactive RC` label if not an active RC.